requirements.txt
python-dotenv
requests
Jinja2
pandas
beautifulsoup4
wradlib
geopandas
pyproj
shapely
cartopy